load of dependencies objects now is correct. Next step is to test the saving of the contact and all the dependencies

+++ arkhe_contacts/src/main/java/com/esalcido/arkhe/contacts/arkhe_contacts/entities/ContactIdent.java
@@ -35,7 +35,7 @@ public class ContactIdent {
         this.documentName = documentName;
     }
 
-    public int getDocumentId() {
+    public long getDocumentId() {
         return documentId;
     }
