+++ arkhe_contacts/src/main/java/com/esalcido/arkhe/contacts/arkhe_contacts/repositories/ContactIdentRepository.java
@@ -9,7 +9,6 @@ import org.springframework.stereotype.Repository;
  * ContactIdentRepository
  */
 @Repository("contactIdentRepository")
-public interface ContactIdentRepository extends CrudRepository<ContactIdent, Long>{
+public interface ContactIdentRepository extends CrudRepository<ContactIdent, Integer> {
 
-    
 }
